feat(auth): separate auth endpoints based on auth strategy
