+++ app/api/login/route.ts
@@ -1,7 +1,7 @@
 import * as bcrypt from 'bcrypt';
 import { omit } from 'lodash';
 import { NextResponse } from 'next/server';
-import { prisma } from '@/lib'
+import { prisma } from '@/lib';
 
 interface RequestBody {
     email: string;
@@ -9,20 +9,20 @@ interface RequestBody {
 }
 
 export async function POST(request: Request) {
-    const body: RequestBody = await request.json();
-    const userExists = await prisma.user.findFirst({
-        where: {
-            email: body?.email
-        }
-    })
-    if (userExists) {
-        const valid = await bcrypt.compare(
-            body.password,
-            userExists.password,
-        );
-        if (valid) {
-            return NextResponse.json({ data: omit(userExists, 'password') });
-        };
-    }
-    return NextResponse.json({ data: null });
+  const body: RequestBody = await request.json();
+  const userExists = await prisma.user.findFirst({
+    where: {
+      email: body?.email,
+    },
+  });
+  if (userExists) {
+    const valid = await bcrypt.compare(
+      body.password,
+      userExists.password,
+    );
+    if (valid) {
+      return NextResponse.json({ data: omit(userExists, 'password') });
+    };
+  }
+  return NextResponse.json({ data: null });
 }
